refactor: rename file names for sign in/up actions.
Files were mistakenly named using camel case. Renamed using kebab case.
